+++ R/zscore.R
@@ -1,8 +1,9 @@
 #' Wrapper for z-normalization
 #'
-#' Wrapper for function \code{\link[base]{scale}} that returns zeros instead of \code{NaN}.
+#' Wrapper for function \code{\link[base]{scale}} that returns zeros instead of \code{NaN}. It also
+#' supports a list of vectors.
 #'
-#' @param x Data to normalize.
+#' @param x Data to normalize. Either a vector or a list of vectors.
 #' @param ... Further arguments to pass to \code{\link[base]{scale}}.
 #'
 #' @return Normalized data.
@@ -10,9 +11,23 @@
 #' @export
 
 zscore <- function(x, ...) {
-     x <- scale(x, ...)
-     x[is.nan(x)] <- 0
-     dim(x) <- NULL # scale returns columns
+
+     if (is.list(x)) {
+          dots <- list(...)
+
+          x <- lapply(x, function(xx) {
+               xx <- do.call("scale", c(list(x=xx), dots))
+               dim(xx) <- NULL # scale returns columns
+               xx[is.nan(xx)] <- 0
+
+               xx
+          })
+
+     } else {
+          x <- scale(x, ...)
+          dim(x) <- NULL # scale returns columns
+          x[is.nan(x)] <- 0
+     }
 
      x
 }